Predelivery tutorial

+++ project/app/src/main/java/edu/ub/pis2019/pis_16/tempac/Presenter/MainMenuPresenter.kt
@@ -3,15 +3,20 @@ package edu.ub.pis2019.pis_16.tempac.Presenter
 import android.content.Intent
 import android.support.v7.app.AppCompatActivity
 import android.widget.Button
+import android.widget.Toast
 import edu.ub.pis2019.pis_16.tempac.Model.HomeWatcher
 import edu.ub.pis2019.pis_16.tempac.Model.MusicService
 import edu.ub.pis2019.pis_16.tempac.Model.OnHomePressedListener
+import edu.ub.pis2019.pis_16.tempac.Model.User
 import edu.ub.pis2019.pis_16.tempac.R
 import edu.ub.pis2019.pis_16.tempac.View.GameActivity
 import edu.ub.pis2019.pis_16.tempac.View.HighScoreActivity
 import edu.ub.pis2019.pis_16.tempac.View.SettingsActivity
+import edu.ub.pis2019.pis_16.tempac.View.TutorialActivity
 
 class MainMenuPresenter(val activity: AppCompatActivity) : Presenter {
+
+    private lateinit var app : TempacApplication
 
     override fun onResume() {
         MusicService.resumeMusic()
@@ -44,10 +49,36 @@ class MainMenuPresenter(val activity: AppCompatActivity) : Presenter {
         })
         mHomeWatcher.startWatch()
 
+        app = activity.application as TempacApplication
+
         activity.findViewById<Button>(R.id.btn_play).setOnClickListener {
-            MusicService.buttonSoundPlay(this.activity)
-            MusicService.destroyReproducer()
-            changeActivityPlay()
+            //we check if the user is in the app.
+            //if(app.loadLocalUser()){
+                if(app.user.fistTry){
+                    app.user.fistTry = false //we set the variable to false to skip the tutorial the second time
+
+                    //changing activity
+                    MusicService.buttonSoundPlay(this.activity)
+                    changeTutorial()
+                }else{
+
+                    //changing activity
+                    MusicService.buttonSoundPlay(this.activity)
+                    MusicService.destroyReproducer()
+                    changeActivityPlay()
+                }
+            /*}else{ //user error, this should never happen xD
+                Toast.makeText(
+                    activity,
+                    "Problem with the user; the score will not be saved",
+                    Toast.LENGTH_LONG
+                ).show()
+                MusicService.buttonSoundPlay(this.activity)
+                MusicService.destroyReproducer()
+                changeActivityPlay()
+            }*/
+
+
         }
 
         activity.findViewById<Button>(R.id.btn_highscores).setOnClickListener {
@@ -73,6 +104,11 @@ class MainMenuPresenter(val activity: AppCompatActivity) : Presenter {
 
     private fun changeActivityHighscores(){
         val intent = Intent(this.activity, HighScoreActivity::class.java)
+        activity.startActivity(intent)
+    }
+
+    private fun changeTutorial(){
+        val intent = Intent(activity, TutorialActivity::class.java)
         activity.startActivity(intent)
     }
 }
